refactor: move generator tree to binding
