+++ packages/context/src/factory/requestContainer.ts
@@ -23,9 +23,7 @@ export class RequestContainer extends Container {
     const definition = this.applicationContext.registry.getDefinition(identifier);
     if (definition && definition.isRequestScope()) {
       // create object from applicationContext definition for requestScope
-      const obj = this.resolverFactory.create(definition, args);
-      this.attachDependencyMap();
-      return obj;
+      return this.resolverFactory.create(definition, args);
     }
 
     if (this.parent) {
@@ -44,21 +42,11 @@ export class RequestContainer extends Container {
     const definition = this.applicationContext.registry.getDefinition(identifier);
     if (definition && definition.isRequestScope()) {
       // create object from applicationContext definition for requestScope
-      const obj = await this.resolverFactory.createAsync(definition, args);
-      this.attachDependencyMap();
-      return obj;
+      return await this.resolverFactory.createAsync(definition, args);
     }
 
     if (this.parent) {
       return await this.parent.getAsync<T>(identifier, args);
-    }
-  }
-
-  protected attachDependencyMap() {
-    for(let [key, value] of this.dependencyMap.entries()) {
-      if(!this.applicationContext.dependencyMap.has(key)) {
-        this.applicationContext.dependencyMap.set(key, value);
-      }
     }
   }
